Create Person constructor with name and age arguments, set those to the Person's name and age. Create an empty stomach array.
Create an 'eat' method that adds 'someFood' to the array with .push();
but ONLY and EXPLICITLY if there are NOT 10 items, (as the prompt says)
"The `eat` method should have no effect if there are 10 items in the
`stomach`"

Create a 'poop' method that empties the stomach.

Create 'toString' method that just returns name and age in a single
string.